gemini/requirements.txt
pandas
python-binance
numpy
matplotlib
